Crate breeze_backend

Modules

dummy
Dummy backend that does nothing.

input
Input handling and traits

ppu

viewport
Provides a utility method for calculating native viewport size when the window is resized.

Enums

BackendAction
An action that can be performed by the user, is detected by the backend and executed by the emulator core.

Traits

AudioSink
Trait for audio backends. Provides methods for writing to a stereo audio channel.

Renderer
Trait for screen renderers. Once per frame, they are given the raw screen data produced by the PPU and can then render this content in a backend-specific way.

Type Definitions

BackendResult
Result with an erased error type.