Struct breeze_backend::dummy::DummyRenderer

pub struct DummyRenderer {
    // some fields omitted
}

Renderer that just does nothing, apart from saving the PPU output for later use. This allows users to extract single rendered frames without having to implement Renderer.

Methods

impl DummyRenderer

fn last_frame(&self) -> &[u8]