Struct breeze_backend::input::joypad::JoypadState

pub struct JoypadState(_);

State of a SNES joypad.

Bits (HIGH | LOW, returned on Data1 from high to low, or left to right): B Y Select Start Up Down Left Right | A X L R 0 0 0 0

Methods

impl JoypadState

pub fn new() -> Self

Creates a new InputState with no buttons pressed

pub fn set(&mut self, button: JoypadButton, pressed: bool) -> &mut Self

Set a button's state

pub fn read_bit(&mut self) -> bool

Reads a bit from the state, as if the state would be stored inside the joypads shift register. This shifts the state to the left and inserts a 1-bit at the right side.